Solution to Calculate and Print the Average of Three Numbers

+++ Calculate and Print the Average of Three Numbers@@6788d5012db25efec946f944/code.c
@@ -1,9 +1,9 @@
 #include <stdio.h>
  int main() {
     int a,b,c;
-    int avg=(a+b+c)/3;
+    float  avg=(a+b+c)/3;
     scanf("%d %d %d",&a,&b,&c);
-    printf("Average: %d",avg);
+    printf("Average: %f",avg);
 
     return 0;
